+++ src/strapi_kit/models/__init__.py
@@ -8,7 +8,16 @@ from .config import RetryConfig, StrapiConfig
 from .content_type import ComponentListItem, ContentTypeListItem
 from .content_type import ContentTypeInfo as CTBContentTypeInfo
 from .content_type import ContentTypeSchema as CTBContentTypeSchema
-from .enums import FilterOperator, PublicationState, SortDirection
+from .enums import (
+    DocumentAction,
+    DocumentStatus,
+    FilterOperator,
+    HttpMethod,
+    PublicationFilter,
+    PublicationState,
+    QueryParam,
+    SortDirection,
+)
 from .export_format import (
     ExportData,
     ExportedEntity,
@@ -61,6 +70,11 @@ __all__ = [
     "FilterOperator",
     "SortDirection",
     "PublicationState",
+    "DocumentStatus",
+    "PublicationFilter",
+    "DocumentAction",
+    "QueryParam",
+    "HttpMethod",
     # Request models - Filters
     "FilterBuilder",
     "FilterCondition",